+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+
+## Added
+* CLDF dataset creation
+* `senses.csv`
+
 ## [0.0.3] -- 2022-10-22
 
 ## Fixed